Add abilty to disable filters.
Most of the changes are to pass along a name with a filter when adding a
filter via add(). This name can be used for a new method
sinsp_evttype_filter::enable(), which takes a name and bool enabled
flag.

To manage the new name and enabled state along with the filter, wrap the
existing sinsp_filter object in a wrapper filter_wrapper containing the
sinsp_filter, enabled flag, and name. Everywhere a sinsp_filter object
was used, a filter_wrapper is used instead.

The list of filters is now a map from name to filter wrapper.

enable just looks up the wrapper and toggles the enabled flag.

In run(), when looping over filters check the enabled flag before
running the filter.